requirements/py_requirements.txt
Django==1.7.2
MySQL-python==1.2.5
django-stronghold==0.2.6
requests==2.7.0
pyquery==1.2.9
python-dateutil==2.4.2
six==1.9.0
django-bootstrap-form==3.2
djangorestframework==3.2.3
markdown
django-filter0.11.0